python/requirements.txt
protobuf==4.25.2
pyserial
